Initial commit.
This changeset implements a very simple linux command that sends
a ping and then parses the response (to extract the ping timings).

It only works with localhost, but it's a start. I first wanted to
implement it through raw sockets, but learned that icmp needs
root access to grab the icmp responses so its easier to just
parse the ping command output which is available on every machine
anyways.